add RBAC

+++ routes/api/beds.php
@@ -6,9 +6,9 @@ use Illuminate\Support\Facades\Route;
 Route::middleware(['auth:sanctum'])->prefix('beds')
     ->controller(BedController::class)
     ->group(function (): void {
-        Route::get('/', 'list');           // list beds
-        Route::post('/', 'store');         // create bed
-        Route::get('/{pid}', 'view');      // view bed
-        Route::put('/{pid}', 'update');    // update bed
-        Route::delete('/{pid}', 'delete'); // delete bed
+        Route::get('/', 'list')->middleware('permission:beds,view');           // list beds
+        Route::post('/', 'store')->middleware('permission:beds,create');       // create bed
+        Route::get('/{pid}', 'view')->middleware('permission:beds,view');      // view bed
+        Route::put('/{pid}', 'update')->middleware('permission:beds,update');  // update bed
+        Route::delete('/{pid}', 'delete')->middleware('permission:beds,delete'); // delete bed
     });
