+++ routes/api/buildings.php
@@ -6,9 +6,9 @@ use Illuminate\Support\Facades\Route;
 Route::middleware(['auth:sanctum'])->prefix('buildings')
     ->controller(BuildingController::class)
     ->group(function (): void {
-        Route::get('/', 'list');           // list buildings
-        Route::post('/', 'store');         // create building
-        Route::get('/{pid}', 'view');      // view building
-        Route::put('/{pid}', 'update');    // update building
-        Route::delete('/{pid}', 'delete'); // delete building
+        Route::get('/', 'list')->middleware('permission:buildings,view');           // list buildings
+        Route::post('/', 'store')->middleware('permission:buildings,create');       // create building
+        Route::get('/{pid}', 'view')->middleware('permission:buildings,view');      // view building
+        Route::put('/{pid}', 'update')->middleware('permission:buildings,update');  // update building
+        Route::delete('/{pid}', 'delete')->middleware('permission:buildings,delete'); // delete building
     });
