+++ routes/api/historyAndPhysicalExaminationFormTwo.php
@@ -6,9 +6,10 @@ use Illuminate\Support\Facades\Route;
 Route::middleware(['auth:sanctum'])->prefix('history-and-physical-examination-form-two')
     ->controller(HistoryAndPhysicalExaminationFormTwoController::class)
     ->group(function (): void {
-        Route::get('/', 'list');          // list history and physical examination form two entries
-        Route::post('/', 'store');         // create entry
-        Route::get('/{pid}', 'view');          // view entry
-        Route::put('/{pid}', 'update');    // update entry
-        Route::delete('/{pid}', 'delete');  // delete entry
+        $module = 'history-and-physical-examination-form-two';
+        Route::get('/', 'list')->middleware("permission:$module,view");          // list history and physical examination form two entries
+        Route::post('/', 'store')->middleware("permission:$module,create");       // create entry
+        Route::get('/{pid}', 'view')->middleware("permission:$module,view");          // view entry
+        Route::put('/{pid}', 'update')->middleware("permission:$module,update");    // update entry
+        Route::delete('/{pid}', 'delete')->middleware("permission:$module,delete");  // delete entry
     });
