+++ routes/api/soaps.php
@@ -6,9 +6,9 @@ use Illuminate\Support\Facades\Route;
 Route::middleware(['auth:sanctum'])->prefix('soaps')
     ->controller(SoapController::class)
     ->group(function (): void {
-        Route::get('/', 'list');           // list soap notes
-        Route::post('/', 'store');         // create soap note
-        Route::get('/{pid}', 'view');      // view soap note
-        Route::put('/{pid}', 'update');    // update soap note
-        Route::delete('/{pid}', 'delete'); // delete soap note
+        Route::get('/', 'list')->middleware('permission:soaps,view');           // list soap notes
+        Route::post('/', 'store')->middleware('permission:soaps,create');         // create soap note
+        Route::get('/{pid}', 'view')->middleware('permission:soaps,view');      // view soap note
+        Route::put('/{pid}', 'update')->middleware('permission:soaps,update');    // update soap note
+        Route::delete('/{pid}', 'delete')->middleware('permission:soaps,delete'); // delete soap note
     });
