+++ routes/api/user.php
@@ -1,17 +1,14 @@
 <?php
 
 use App\Http\Controllers\UserController;
-use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 
 Route::prefix('user')->controller(UserController::class)->group(function () {
-    Route::post('/register', [UserController::class, 'register'])->middleware(["auth:sanctum"]);
-    Route::get('/list', 'list');
+    Route::post('/register', 'register')->middleware(["auth:sanctum", "permission:users,create"]);
+    Route::get('/list', 'list')->middleware(["auth:sanctum", "permission:users,view"]);
     Route::post('/logout', 'logout')->middleware(["auth:sanctum"]);
-    Route::put('/{pid}', 'update');    // update note
-    Route::get('/{pid}', 'view');          // view note
-    Route::delete('/{pid}', 'delete')->middleware(["auth:sanctum"]);
+    Route::put('/{pid}', 'update')->middleware(["auth:sanctum", "permission:users,update"]);
+    Route::get('/{pid}', 'view')->middleware(["auth:sanctum", "permission:users,view"]);
+    Route::delete('/{pid}', 'delete')->middleware(["auth:sanctum", "permission:users,delete"]);
 });
-Route::get('/user', function (Request $request) {
-    return $request->user();
-})->middleware('auth:sanctum');
+Route::get('/user', [UserController::class, 'me'])->middleware('auth:sanctum');
