+++ routes/api/vitalSigns.php
@@ -6,9 +6,9 @@ use Illuminate\Support\Facades\Route;
 Route::middleware(['auth:sanctum'])->prefix('vital-signs')
     ->controller(VitalSignController::class)
     ->group(function (): void {
-        Route::get('/', 'list');           // list vital signs
-        Route::post('/', 'store');         // create vital sign
-        Route::get('/{pid}', 'view');      // view vital sign
-        Route::put('/{pid}', 'update');    // update vital sign
-        Route::delete('/{pid}', 'delete'); // delete vital sign
+        Route::get('/', 'list')->middleware('permission:vital-signs,view');           // list vital signs
+        Route::post('/', 'store')->middleware('permission:vital-signs,create');         // create vital sign
+        Route::get('/{pid}', 'view')->middleware('permission:vital-signs,view');      // view vital sign
+        Route::put('/{pid}', 'update')->middleware('permission:vital-signs,update');    // update vital sign
+        Route::delete('/{pid}', 'delete')->middleware('permission:vital-signs,delete'); // delete vital sign
     });
