+++ routes/api/wards.php
@@ -6,9 +6,9 @@ use Illuminate\Support\Facades\Route;
 Route::middleware(['auth:sanctum'])->prefix('wards')
     ->controller(WardController::class)
     ->group(function (): void {
-        Route::get('/', 'list');           // list wards
-        Route::post('/', 'store');         // create ward
-        Route::get('/{pid}', 'view');      // view ward
-        Route::put('/{pid}', 'update');    // update ward
-        Route::delete('/{pid}', 'delete'); // delete ward
+        Route::get('/', 'list')->middleware('permission:wards,view');           // list wards
+        Route::post('/', 'store')->middleware('permission:wards,create');         // create ward
+        Route::get('/{pid}', 'view')->middleware('permission:wards,view');      // view ward
+        Route::put('/{pid}', 'update')->middleware('permission:wards,update');    // update ward
+        Route::delete('/{pid}', 'delete')->middleware('permission:wards,delete'); // delete ward
     });
